feat: 界园肉鸽不期而遇实现选项识别 (#14540)
* feat: RoguelikeEncounterOptionAnalyzer

* feat: 界园肉鸽实验性功能 -- 识别选项数量后调整选项

* chore: Auto update by pre-commit hooks [skip changelog]

* refactor: 跳过不可用选项

* refactor: 设置初次滑动

* refactor: 给 save_img 日志加点描述

* feat: 加个 get_img 接口，以便于获得合成图

* chore: Auto update by pre-commit hooks [skip changelog]

* fix: 不是你怎么没有 formatter 啊？

* refactor: 把分析数据存储在插件成员变量内，为未来重构打基础

* refactor: 把不期而遇选项识别结果输出移到插件中，加入 callback

* feat: 无法识别选项文字时保存图片

* refactor: 给 save_img 日志加点描述

* refactor: int -> size_t

* refactor: choose -> select

* refactor: sort functions

refacor: sort functions

* refactor: 用 option_analyzed 来判断是否已经识别过不期而遇选项

* fix: 去掉不必要的参数 theme

* refactor: 给 OCR 配置任务加点东西（大概率没什么用吧）

* refactor: 去掉内部接口中的冗余 sanity check

* feat: 加入 view 控制

* fix: 忘了 callback 不能用 const

* feat: 处理一下 edge case

* feat: 来点 WpfGui Log

* fix: typo

* fix: 一些漏网之鱼

* fix: 再调整下滑动距离

* fix: 什么乱七八糟的

* fix: 什么乱七八糟的

* fix: 方向错了，加点 Log

* fix: 修正长图识别 roi

* fix: 再调整一下 roi

* fix: 所以说不建议再睡眠不好的时候糊屎

* fix: 对不可选选项的模版进行二值化后再 OCR

* fix: 好像得考虑一下通道

* fix: 删掉无用参数

* refactor: 再整理一下代码，减少冗余

* chore: Auto update by pre-commit hooks [skip changelog]

* fix: 再缩小点 roi

* fix: 加点延迟

* fix: clang-tidy 说这样写

* feat: 把 debug 截图加回来，收工

* docs: 给任务链加点注释

* docs: 给 RoguelikeEncounterOptionAnalyzer 加点注释

---------

+++ src/MaaCore/Config/Roguelike/RoguelikeStageEncounterConfig.h
@@ -51,12 +51,12 @@ public:
     struct RoguelikeEvent
     {
         std::string name;
-        int option_num = 0;
-        int default_choose = 0;
+        size_t option_num = 0;
+        size_t default_choose = 0;
         std::vector<ChoiceRequire> choice_require;
         std::string next_event;
 
-        std::vector<std::pair<int, int>>
+        std::vector<std::pair<size_t, size_t>>
             fallback_choices; // 备用选项，格式为 (选项数量，选择的选项)，有些事件的选项数量可变
     };
